Xilsecure:Implicit conversion of underlying type (MR).
This patch fixes below MISRA-C violation:
Implicit conversion of underlying type (MR).
MISRA-C:2012/AMD1 R.10.3,R.10.4


Acked-for-series: VNSL Durga Challa <vnsl.durga.challa@xilinx.com>

+++ lib/sw_services/xilsecure/src/zynqmp/xsecure.c
@@ -948,7 +948,7 @@ UPDATE:
 		IsEncrypted = 1U;
 	}
 
-	EncOnly = XSecure_IsEncOnlyEnabled();
+	EncOnly = (u8)XSecure_IsEncOnlyEnabled();
 	if ((EncOnly != 0U) && (IsEncrypted == 0U)) {
 		Status = XSECURE_ENC_ISCOMPULSORY;
 		goto END;
